DefaultSerializer supports str types only.
Behavior was not set explicitly although it was the original intention.
Due to incompatibilities between memcached, redis and memory, DefaultSerializer
was able to store different types depending on the backend. This is
unwanted incompatibility between backends and thus, has been restricted
to avoid surprises
