+++ tests/acceptance/conftest.py
@@ -60,3 +60,12 @@ def memcached_cache(event_loop):
 
     event_loop.run_until_complete(cache.delete(pytest.KEY))
     event_loop.run_until_complete(cache.delete(pytest.KEY_1))
+
+
+@pytest.fixture(params=[
+    'redis_cache',
+    'memory_cache',
+    'memcached_cache',
+])
+def cache(request):
+    return request.getfuncargvalue(request.param)
